20 rename package v2 (#25)
* feat: change root directory

* refacto: rename module and fix creation file error
