+++ root.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/Damien-Venant/prev-updater/cmd"
+
+func main() {
+	cmd.Execute()
+}
